Add duplicate service with correct LGSL code
What

We want to add a duplicate record for the lateral flow mass testing
finder service with a code of 1827. Once the
local_transactions:fetch_and_clean rake task has been run - the version
of the service with the LGSL code of 100001 will be removed and the rake
task re-run. It is hoped that this will keep the service running
smoothly without interruption.

Why

When the lateral flow mass testing finder service was created, it was
created with a dummy code of 100001.  We now have a real code (1827) so
would like to use this code instead of the dummy one.