Update LocalService LGSL code
What

Add a LocalService with an LGSL code 1827.

Why

The service was created using a dummy LGSL code (100001).  We now have
a real LGSL code for this service (1827) and would like to use it
instead of the dummy code.

For safety and to ensure there is no interruption to the service we need
to add the 1827 service - leaving the 100001 service in-place - until
the switch is complete. At that point we can remove the 100001 service.