Fix Style/OptionalBooleanParameter offences
This rule demands that boolean arguments are passed in as keyword
arguments for enhanced readability. In two cases I found that the
arguments weren't actually used and could be removed. In another I
switched to a keyword argument. Finally for the UpdateWorker, we
couldn't use a keyword argument as Sidekiq doesn't accept them so I
ignored the rule.