Fix Lint/ConstantDefinitionInBlock offences
This resolves a number of constants that were defined to global scope.
Most of these could be resolved by switching the constants to variables,
however I couldn't do this for the EditionPresenterFactoryTest as that
needs the constant to be globally available. I wasn't able to work out
how to stub a constant with the mocking library for this code so I
defined the constant at the class scope.