Add Edition#paths method
This returns the paths associated with a particular edition, so that
includes the base path and any parts.

We might want to think about using this method to make the routes more
explicit when we put the content into the Publishing API, with `exact`
routes rather than `prefix` ones:
https://github.com/alphagov/publisher/blob/742541958fb591943dadf1c3bd2e52dbf2811491/app/presenters/formats/edition_format_presenter.rb#L67-L71

For now, this method will be used to generate a sitemap from just
Mainstream Publisher.