Add SitemapController
This generates a sitemap.xml file containing all the paths published by
Mainstream Publisher. This has been requested as the content team would
like to start using a tool called Siteimprove which will consume this
sitemap.

I've intentionally not used the standard /sitemap.xml route, as it's not
a sitemap of the Publisher app, it's a sitemap of the pages published by
the app.

It's open to the public so that Siteimprove can access it, but we may
need to re-think this in the future if it's starts getting indexed by
other search engines. The work with Siteimprove is only going to be
running for a year, so we might be able to remove it after that.