Duplicate DA values for new editions
Devolved administration availability modelling was added in a [previous PR](#1456).

A bug was discovered, in that the embedded fields in DevolvedAdministrationAvailability were not being copied when a new edition was created.

The reason for this is the scotland_availability, wales_availability and northern_ireland availability fields were not included in the array returned by the[type_specific_field_keys](https://github.com/alphagov/publisher/blob/main/app/models/edition.rb#L499). The test for duplicating editions used a different edition type, which is partly why the bug was missed.

This commit overrides the edition#build_clone method in the child class to clone the devolved administration availability fields when a new edition is created.

Co-authored-by: Kevin Dew <kevin.dew@digital.cabinet-office.gov.uk>