Add devolved admin specific attributes to the strong params hash
Add the devolved administration attributes (type and alternative_url) so
that the user can save and view previously saved values in the user
interface.

Currently, these attributes are not being saved when changed by the 
user.

We are using some prior-art from the part.rb and parted.rb models which
ensure that the strong params hash has the correct keys and values when
being saved and viewed via the local transaction edition form.

Without '_attributes' being appended to the param name and the
'accepts_nested_attributes_for' call, the controller is unable to save
the record as some of the devolved admin attributes are undefined.