Add devolved admin specific attributes to the user interface
Add the devolved administration attributes (type and alternative_url) to
the existing local transaction edition form.

This will allow the user to view - and ultimately update and change
these attributes via the user interface.

We add various styles to these fields so that they match the required
appearance of the proposed design.