Add Rake task to delete old Artefact reserving needed slug
This Artefact is an old StartPage that has not been touched since 2015.
StartPages are now published in SmartAnswers, and this example is
leftover from before they were moved from Publisher.

There is now a need for Publisher to re-use the `landlord-immigration-check`
slug, and this new Artefact cannot be created while the slug is associated
with the old StartPage.