Merge associated model validation errors
Where an Edition has embedded associations (for example:
LocalTransactionEdition and SimpleSmartAnswerEdition), the default
message that is currently shown for embedded model validation error
messages is a rather unhelpful "is invalid".

We want the embedded model validation error messages to be more
meaningful and helpful to the user.

To do this we need to merge the embedded association messages into the
parent's errors array so that they are easier to display and (in the
case of Ajax responses), easier to place in context with the appropriate
field. This process needs to be repeated if an embedded association
itself has an embedded association.

Here are a couple of before-and-after examples...

LocalTransactionEdition

before = [
  "Scotland availability is invalid",
  "Wales availability is invalid"
]

after = [
  "Scotland availability [\"Alternative URL must be a full URL,
  starting with https://\"]",
  "Wales availability [\"Alternative URL can't be blank\"]"
]

SimpleSmartAnswer

before = [
  "Nodes is invalid"
]

after = [
  "Question 1 [\"Title can't be blank\", \"Label can't be blank\"]",
  "Outcome 1 [\"Title can't be blank\"]"
]