Merge devolved nation availability validation errors into parent class
The LocalTransactionEdition class has a one-to-one relationship with the
Scottish, Welsh, and Northern Irish devolved nation availability
classes that indicate whether or not each devolved nation has their own
webpage for the local transaction service, use the general UK-wide page,
or do not offer the service at all.

These child classes generate their own model validation errors which
need to be made visible to the user. These errors generally revolve
around the URL, which cannot be blank and must be a valid HTTPS address.

Currently, when a model validation fails the user is shown a rather
unhelpful "is invalid" message.

We want the user to be given more specific - field level - validation
error messages to so that they are more meaningful and helpful to the
user when diagnosing and correcting the issue.

To do this we need to remove the unhelpful "is invalid" message and
merge the devolved nation availability validation messages into the
errors array so that they are easier to display and easier to place in
context with the appropriate field.