Update the way model validation errors are presented to the user
Now that we have a form helper in place, we need to update the view so
that we can display the validation messages as close as possible to the
field and highlight the field and label.

The application updates records in 2 ways: HTML form and Ajax-based POST
request. So we need to handle validation errors being returned from
either of these methods. We also need to handle multiple errors for a
single field.

Tests are updated as appropriate.