Create a form helper class
We want to be able to display validation error messages in context
(physically as close as possible to the field in question) to enhance
the user experience and assist the user in diagnosing and correcting the
issue.

We also want to minimise the amount of boilerplate code and create
consistency of the HTML structure used on all fields that are subject to
a model validations. To do this we wrap the field in a 'form-group'
which requires the outer helper method to yield to the actual input tag.

We also introduce an "error-block" in which to display the validation
errors.

It is assumed that fields may have more than one validation message, so
messages are displayed as a list. These methods handle any optional
elements required by the structure surrounding the field.

Within the tests, we stub the form 'object' - providing it with an
'errors' hash that is initialised to 'Hash.new([])'. This creates a Hash
where both the keys and values are an empty array, which protects us
nil's being returned from method calls and saves us from having to add
guard clauses.

Finally, we update a couple of tests that pass in a `Hash` instead of an
`ActiveModel::Errors` object by adding the expected error into an
`ActiveModel::Errors` instance so that it satisfies what the code
expects.