Update edition reviewer field HTML structure
The reviewer field does not have model-level validation against it.
However, the business logic of the field (e.g. the Reviewer cannot also
be the Assignee) is validated by the ReviewerValidator class.

Because of this we want to apply the same validation message display,
and HTML structure as all other model-level validation fields for
consistency. We use the newly created `FormHelper` class to achieve
this.