Update 'save_edition_and_assert_error' method to take a error string
Currently, the 'save_edition_and_assert_error' test integration helper
only checks for basic text on the page - "We had some problems saving".
This text is present on the page - in both the HTML form response and
the Ajax POST response - whenever model validation error is generated.

It would be nice if it were to also take an optional - more specific -
error message and assert that this error text is also on the page.

Update existing tests to use the new optional error message param.