Add a Variant error block for without JavaScript requests
TransactionEditions can be saved both with and without JavaScript
enabled.

When saved with JavaScript, Variant model validation errors are
dynamically appended to the appropriate field (via JavaScript) from the
response JSON.

However, when saved without JavaScript, Variant model validation errors
are missing. This is because the page is reloaded and there is nowhere
to render or show the errors.

This commit renders any Variant model validation errors when saving a
TransactionEdition without JavaScript.