Fix SimpleSmartAnswer Node and Options UI issues
Currently, model validation error messages are not being shown for the
Node title and the Option label fields. Also, the UI breaks when model
validation errors are shown for the Option next_node field.

This does not help the user identify and resolve model validation errors
quickly and easily and is generally a sub-optimal user experience for
the user.

This commit:

* Adds model validation error display (when generated) for the Node
title field.

* Adds model validation error display (when generated) for the Option
label field.

* Moves the existing model validation error message display for the
Option next_node field as this was breaking the UI when present.

* Brings the HTML structure of the model validation displays above
in-line with the structure introduced in [this PR](https://github.com/alphagov/publisher/pull/1478).

* Applies a row a grid layout from the pre-existing Bootstrap framework
to each Option field group. This improves the user experience both on
the desktop and mobile/tablet devices.