Tweak the styling of the SimpleSmartAnswer Node Options UI
The display of the Options UI does not respond well when the user is
using a tablet or mobile device. In particular the fixed-width next_node
field causes significant breaks in the UI and diminishes the user
experience.

We should make the UI: as responsive as possible across device types, as
consistent in use as possible, and provide an improved experience for
the user.

This commit updates the display of the Node Options input and select
fields so that they now take up the entire gird space they are in and
ensures all Option fields are vertically aligned to the top of their 
container.