Bump select2-rails from 3.5.9.1 to 3.5.11
Bumping `select2-rails` to the latest patch of version 3.5 to remove the
requirement for `thor (~> 0.14)`, which has required updates to the
tests as the `select` element is explicitly marked as `display: none`
(so not matched with capybara's `select` method).

We cannot upgrade to version 4 without a substantial amount of front end
work, as the styling and marking up of elements has changed
significantly in this version.

Instead bumping to 3.5.11 to remove the dependency on an older version
of `thor`, which was blocking the upgrade to Rails 6.1.