Use capybara-select-2 for select tests
We are using the capybara `select` method to choose items from dropdown
menus in tests. This method only finds visible elements.

In version 3.5.10 of `select2-rails`, the `select` element has an
explicit `display: none` styling applied [1] where the dropbox is rendered
as a `select2`, therefore capybara can no longer find the element.

There is an unused `select2` method in the test helpers, but this
appears not to function correctly or be used.

Adding `capybara-select-2` gives us the `select2` selector, which will
allow us to test this functionality once again and allow us to remove
the `select2` method.

1: https://github.com/argerim/select2-rails/compare/v3.5.9.1...v3.5.10#diff-ffea88f08f5750e44aab693b050ad1f4e03fc54ec95a48f814dd67395ec0df95R886