Add missing label for reviewer field
In d356608, the structure of the reviewer field partial was changed. In
doing so, it appears the label was removed.

By default, we appear to be humanizing the form group name (i.e.
`reviewer` to `Reviewer`). Adding the name back in so it is explicit
where the label is coming from.