Remove js-error CSS class
The js-error CSS class is appended to elements when model validation error messages are returned from an Ajax POST/Update request on an Edition.

The has-error CSS class is appended to elements when model validation errors messages are returned from a standard Rails HTTP form POST request on an Edition.

It is somewhat unnecessary to have two different classes which do the same thing so we are just going to utilise the 'has-error’ class included with bootstrap to meet our requirements.

There was some debate as to whether we should define our own CSS for error messaging; the rationale being that it would not be affected by future updates to bootstrap. We decided against doing that for now.

In addition to replacing 'js-error' we have removed some redundant code in the tests, namely the first two lines of the 'it shows the error messages' function.

We also the  'removes validation errors on the newly added part' test more specific by pointing it directly at part 4, instead of scanning the whole page for the element.