Fix html structure for errors
When testing locally we noticed that the error messages were not showing where we expected them to be, so we have shifted them slightly to fix this.

Another interesting change was changing the 'removeValidationMessages' function to use empty() as opposed to remove(). Empty removes all the child nodes of the selected elements instead of just flat out destroying the element, this is desirable for our purposes.

One quirk of this change is that it resulted in our 'simulation of dragging' test to fail, namely because now there was now an empty div where there had not been previously. To fix this we changed the 'tolerance' of the simulateDragSortable function. Tolerance determines the number of pixels the simulated drag can overlap on an element height, the default is 50% but we have changed it to 0 so that there is no ambiguity. This fixes the tests as we don't have to worry about the height of elements.