Add a view partial for the common Part and Variant attributes
The Part and Variant models both have title and slug attributes.

However, there are differences between the HTML currently used for these
input fields in the Part and Variant view partials. Notably, the errors
list is within the form-wrapper for Variants and outside for Parts.

This change creates a common view partial for use by all formats that
require Parts or Variants. It calls specific Part or Variant partials
which render the model-specific attributes. 

To facilitate this we pass in a new variable - child_record_type - which
is currently either 'part' or 'variant'. This is also used where the 
on-screen wording needs to vary.