Update the Variant view partial to display only Variant attributes
We now have a view partial that displays the attributes common to both
Parts and Variants. 

This common partial in turn renders either the Part or Variant specific
partial. So, the Variant partial only needs to render attributes that
are specific to Variants.

We also rename the partial to bring it in-line with the naming
convention used by the Part partial, and so that we can render the
Variant partial - in the common_part_attributes partial - by simply
using the child_record_type (in this case 'variant') that is passed in
as a parameter.