Pass the new date select fields into the downtime_params
This sucks, but as far as i can see there's no way to handle this
more gracefully without introducing something like a form object.

As the `select_x` methods each require a field name, i can't associate
the start_time and end_time attrs on the Downtime object.

This means that i have to munge the params in the controller to be able
pass the correct values to the Downtime before it is updated.