Combine smart answers, nodes and editions with errors into one object
This commit adds a method to the controller which gives access to an
array of objects which contain errors. This is needed for smart answers
as the form on the edit publication page contains 3 different models.
SimpleSmartAnswerEdition's, Nodes, and Options (for questions).

This means that the SimpleSmartAnswerEdition has nested objects with
errors. The construct_objects_with_errors pushes all objects with an
error into an array which can be used in the view to render errors.

This method is called in the show action and any other action which
renders the show view on failure to save.