Make validation errors descriptive for smart answers
This ensures that error messages for validation errors are
descriptive. As the page has a lot of questions/outcomes each of which
often has many options. Error messages need to tell the user exactly
where the error has taken place.

This commit constructs error messages based on the ordering of the node
order & option order.