Add the `object.html_ref_for_error`
This adds the `object.html_ref_for_error` to the Edition,
SimpleSmartAnswerEdition, Node and Option models.

The form builder assigns an id to the relevant node and input fields
based on the node order attribute and the ordering of the options.

In order to link the errors in the error-summary to the correct inputs
we have to replicate the construction that the form builder uses to
assign them.

It follows this pattern:

For top level objects (so in this case smartanswers):

"#edition_#{attribute}"

For nodes:

"#edition_nodes_attributes_#{order - 1}_#{attribute}"

The first attribute number is based on the index of the node
simple_smart_answer_edition.nodes array. Hence order - 1.

For Options:

"#edition_nodes_attributes_#{node.order - 1}_options_attributes_#{order}_#{attribute}"

It also needs to be added to the edition subclass so when called on all
other edition types (i.e. guides) it continues to work by defaulting to
"#edition_#{attribute}".