Add an error-summary to the ajaxSave module
This adds the error summary banner to the Edit Publication page when JS
is on. This uses the same design as the non-JS pages.
The correct behaviour for this is the following.

When an error is triggered:

The error summary banner shows. Each error links to the correct inputs.

When the form is resubmitted with errors:

Errors are not present for fields that have been have been completed.

When the for is resubmitted successfully:

The banner disappears.

When a parent edition has the message 'is invalid' or an error is
returned with the errorKey 'parts':

We filter out the errors. This is because 'is invalid' is a summative
error which checks the validity of the nested objects. Due to this it
does not link to an input, which is covered by errors that live of the
nested objects.

Parts errors are filtered out for the same reason. They are handled in
the next commit.