Fix flakey test
The downtimeWithInvalidDates test was flakey when being run before 10am. This is because it selects from an hour dropdown with hours padded with '0' when they are single digits. So 9am is '09'. As the test selects based on the current time, and does not pad the number, when it tries to find, e.g. the string '9' it resorts to fuzzy matching as an exact match is not found, and then hits a Capybara ambiguous error. So this fix pads the number when it is a single digit so that an exact match is found.