Add launchy gem to test dependencies
This is a useful dependency for development workflow while using capybara. Inside a capybara test, you can insert `save_and_open_page` to view the page in a browser, helping to debug tests. Note that this will not work while running tests with govuk-docker as it requires a non-headless browser.