Don't call .class when using case statement to switch based on edition type
Mongoid 7.4 updated the '===' operator to match ruby syntax, which means that it is equivalent to calling .is_a?(Class). This is what the case statement uses under the hood, and therefore the upgrade to Mongoid 7.4 necessitates removing .class from this case statement. For more details, see the release notes: https://www.mongodb.com/docs/mongoid/master/release-notes/mongoid-7.4/#change-operator-to-match-ruby-semantics