Use database cleaner specifically for mongoid
Since `database_cleaner 2.0.0.beta 2020-04-05`'s release the
functionality of the gem has been split into separate gems. We now need
to use the mongoid version of database cleaner in line with the database
we use.

In doing this we have to change the database cleaning strategy to
`:deletion` instead of `:truncation` since deletion is the only
available strategy for mongo/mongoid databases