Use LTS Node version in Dockerfile
This fixes an issue where the command to add nodesource was failing
because `curl` wasn't installed. This meant that the node version
bundled with the image was installed which was 10.24.0
It also updates the version referenced in nodesource to be the LTS node
version.

Rather than installing npm and using that to install yarn it installs
yarn directly.

Finally rather than re-installing node in the runtime image (and ending
up with Node 10.24.0 again) it copies node from the builder image.