Stop reusing test fixtures
The ContentWorkflowPresenter test was reusing Actions between two different Editions.

A change of behaviour between Mongoid 7 and Mongoid 8 meant that the Actions were being removed from the first Edition when they were embedded into the second one.

This resulted in test failures which expected the 5x Actions to become 10 when saved on two Editions.