Fix broken validation test
It looks like Mongoid 8 changes how/where/when typecasting occurs.

So entering a String like "word" into the form now gets typecasted to an empty value, meaning the `presence` validation triggers rather than the `numericality` validation.

To work around this, I've changed the 'bad input' value to be a Float so that it still triggers the `numericality` validation.