Fix deploy GitHub Action
This updates the deploy GitHub Action workflow to correctly build and trigger a
deployment of the application to the EKS. This configuration may have been
incorrect as only subset of app originally were set up correctly. This is a
part of bulk set of PRs raises. The workflow is still configured to be
triggered on_push, until CI workflow has been configured correctly.