Add platforms for precompiled gems
This adds Apple Silicon (arm64-darwin), Linux ARM 64 (aarch64-linux) and
Linux x86 (x86_64-linux) architectures as platforms for gems.

This has been done to resolve an issue where Ruby gets confused about
what gems are available on a system.

The issue we experienced was Ruby scripts thinking they couldn't find
gems with an error of:

```
Could not find mini_portile2-2.8.0 in any of the sources
Run `bundle install` to install missing gems.
```

The reason this gem couldn't be found was because mini_portile was not
installed as there was already a nokogiri 1.13.8.x86_64-linux installed
on the machine and thus didn't need the mini_portile dependency. By
actually supporting the x86_64 platform our gemfile becomes aware that
this dependency is not needed which resolves the confusion.

An alternative way to resolve this would be specify in our deployment
script [1] to force a platform of ruby, however this has the consequence
of making deploys take longer.

This change also means installing precompiled gems like nokogiri will be
faster on devices.

[1]: https://github.com/alphagov/govuk-app-deployment/blob/75c88f0161412455294846fdfc8e4b6d8145b867/recipes/ruby.rb#L90-L93