Update govuk_publishing_components to resolve Plek warning
This resolves the "Plek.current is deprecated and will be removed.
Use Plek.new or Plek.find instead." warning on initialisation.

This gem doesn't get updated automatically by dependabot as it's an
indirect dependency through Govspeak.

As part of this `bundle update` nokogiri is updated, which seems to add
in platform gems that I'd have expected to be there in:
https://github.com/alphagov/publisher/commit/f9418c57751ed50134cf0bf1c59e3b4b062e3633
🤔.