Apply fixes for mail 2.8.0
From my brief digging it seems that the mail gem no longer allows Mail
headers to have a value of nil and will discard the header. These won't
be discarded if the value is an empty string.

The code that this is being tested requires an empty string for the
Return Path heading (so this must previously have relied on some type
casting) whereas just the presence of X-Autorespond and
X-Auto-Response-Suppress is checked and thus the value has no impact.