Include locale in call to publishing api to discard draft edition
There is a bug whereby when an edition with a non-english locale is deleted in Mainstream Publisher,
it attempts to discard the draft from the Publishing API. However, this fails because the 'locale'
option is not provided, so the Publishing API assumes its an english edition. This commit fixes
that by ensuring that the locale is passed to Publishing API for this discard message.