Remove action buttons from retired documents
Previously, retiring a document only disabled creating a brand new
edition and creating a new edition of an existing document. However,
there were a number of POST actions that a user could perform, like
update tagging, change format + saving the document.

If we mark the document type as 'retired' then we don't expect the
inidvidual documents to be modified. All the documents data can be
displayed but not edited + internal notes + unpublish feature (incase a
document needs unpublishing).

Trello:
https://trello.com/c/85Kl4zO7/2092-disable-editing-of-tags-when-marking-publisher-licences-as-retired