Document Simple Smart Answers
A search for "Simple Smart Answers" in the Developer Docs returns nothing relevant to the Simple Smart Answer format:

> ![Screenshot 2023-08-29 at 06 49 51](https://github.com/alphagov/publisher/assets/5111927/fab80b04-86d9-44b0-90fe-29a5df7cd00d)

This commit therefore documents the format. It uses the name of the format for the h1 tag, so should appear at the top of the search results. It's documented close to the code where it is defined, made possible by Developer Doc's "external doc import" mechanism.