Use current user as assignee
Previously, a publisher would need to manually assign themselves as the
assignee when creating an edition. We now assign the user that
created the edition as the assignee. This can be still be modified from
the UI if the assignee is different from the creator, but for the
majority of cases this should be the expected behaviour.

Trello:
https://trello.com/c/3zzF4dKj/37-auto-assign-ownership-in-mainstream-publisher