Display date/time field-level validations
When the 'from' datetime or 'to' datetime fails validation, display the validation error (for both the date and time field groups).

Although they appear as two separate sets of fields on the form—one for the date and one for the time—on the backend model it is a single 'datetime' field. As the validation comes through for either `start_time` or `end_time`, we don't know whether it's the date or time part that is invalid (or both), so display the error message in both field groups.

Since the field groups appear side-by-side, it also looks weird when only the date or time field group has an error message (as the error message pushes down the fields within that field group).

This does mean that the error appears twice on the form, but this is deemed acceptable for now.