Add temp rake task to fix data integrity issues
Somehow energy-company-obligation and manage-your-tax-credits have ended
up with Artefacts with kind "completed_transaction", even though their
editions are Answer and Transaction (note _not_ CompletedTransaction).

This prevents the editions from being published, because the artefacts
are not valid because their slugs don't begin with "done/"

We ran equivalent code to this rake task in integration which appears to
have fixed the issues. We'll test again in staging before running in
production.

https://govuk.zendesk.com/agent/tickets/5819067