Rename RootController to LegacyRootController
This work is part of upgrading Mainstream Publisher to use the GOV.UK Design System, in place of the current, and outdated, bootstrap UI.

This commit renames the `RootController` class to `LegacyRootController`. The next commit will create a new `RootController` class, with a feature toggle to control which is used.

Some tests, which were asserting that the "index" action of the "root" controller was being redirected to, have been updated to instead assert that the `root_path` has been redirected to so that they are agnostic to this change.

Renaming the legacy controller now (rather than creating a new controller for the new version) will make cleanup at the end easier when we remove the feature toggle and all legacy root code.