Push fact check metrics to prometheus
Inspired by https://github.com/alphagov/search-api-v2/pull/209

This adds the prometheus-client gem, and then uses it to send fact check emails metrics to prometheus via the pushgateway.

We can't use the conventional prometheus collector, because this code runs as a script, rather than a web process.

This should not be merged until we've set the PROMETHEUS_PUSHGATEWAY_URL environment variable.