Send `last_edited_by_editor_id` to Publishing API
This is a new field that will allow us to identify the last person who
has edited an Edition within the Publishing API.

When the Edition is the first Edition in a series, this will be the
user who has carried out the `CREATE` action, otherwise it’ll be the
user who has carried out the `NEW_VERSION` action.

I’ve added a `it_includes_last_edited_by_editor_id` test to add to all
the edition-shaped presenter tests to be a bit more DRY and give us
confidence that the information is getting sent correctly.