Remove incorrect updated_at call from public_updated_at
Our previous behaviour was reporting the public_updated_at as the
last time the documented was updated overall. This was causing
the document to be misreported as having major updates every change.

By default the publishing_api handles this by updating the field only if the
change is a major one, so this is an unnecessary override. As the field can
now be nil, it has been made optional.