Prevent updated_at being changed when edition is archived
This should fix most causes of behaviour we are seeing wherein the
public_updated_at field is being set incorrectly in some circumstances,
causing the documents to become visible in various time based search
results and potentially causing confusion for the user.

This function is only called by edition when archiving previous
versions of itself, and the fact check message processor. In order to
keep the expected behaviour of the latter, a new function was written
that provides us with a means to update without affecting the time stamp.