`#is_for_edition?` is false when not published
When an edition is archived, but not published, this was raising an
error, as `edition.publised_at` was nil. This adds an additional check
to see if the edition had ever been published. If not, then we know this
update didn’t occur during the lifecycle of this edition, as it was
never public.