show Content Block update events on in progress Hosts
Previously an event would only show in the
Edition's "history and notes" if the event took
place after the Edition had been published (and if
it was not archived/superseded).

We now want to display event if it took place at
any time after an Edition's creation, IF it is not
archived.

Looking at the db and UI, if an Edition is
"unpublished" then it has an Artefact with state
`archived` and an Action as well. However, if an
Edition's state is `archived` from being superseded,
it does not have the corresponding Artefact OR
Action. So we have to account for those two
scenarios - if an edition has been archived
somehow without the corresponding artefact AND
without being superseded I don't think there is
any reliable way of knowing when it was archived,
so we just return false.